Moving `/metadata` components over to `/record`.
