--- src/modules/metadata/index.js
@@ -1,5 +1,0 @@
-import Metadata from './components/Metadata';
-
-export {
-  Metadata
-};
