+++ src/modules/records/index.js
@@ -15,6 +15,7 @@ import FullRecordPlaceholder from './components/FullRecordPlaceholder';
 import ILLRequestMessage from './components/ILLRequestMessage';
 import KeywordSwitch from './components/KeywordSwitch';
 import MARCTable from './components/MARCTable';
+import Metadata from './components/Metadata';
 import Pagination from './components/Pagination';
 import RecommendedResource from './components/RecommendedResource';
 import Record from './components/Record';
@@ -40,6 +41,7 @@ export {
   loadingHoldings,
   loadingRecords,
   MARCTable,
+  Metadata,
   Pagination,
   RecommendedResource,
   Record,
